+++ lib/system/web/src/lib.rs
@@ -263,7 +263,7 @@ pub fn get_webgl2_context
     context.dyn_into().map_err(|_| no_webgl())
 }
 
-pub fn request_animation_frame(f:&Closure<dyn FnMut(f32)>) -> Result<i32> {
+pub fn request_animation_frame(f:&Closure<dyn FnMut(f64)>) -> Result<i32> {
     let req = window()?.request_animation_frame(f.as_ref().unchecked_ref());
     req.map_err(|_| Error::missing("requestAnimationFrame"))
 }
